PEP8 and clean-up
